+++ src/commands/index.ts
@@ -2,6 +2,7 @@ import * as vscode from 'vscode';
 
 import { registerApplySqlovelyExtrasCommand } from './applyExtras';
 import { registerFormatCurrentFileCommand } from './formatCurrentFile';
+import { registerFormatSqlFilesInDirectoryCommand } from './formatSqlFilesInDirectory';
 import { registerInsertOrUpdateMetadataHeaderCommand } from './insertOrUpdateMetadataHeader';
 import { registerShowActiveDialectCommand } from './showActiveDialect';
 import { registerSwitchDialectCommand } from './switchDialect';
@@ -11,6 +12,7 @@ export function registerCommands(context: vscode.ExtensionContext): void {
     registerShowActiveDialectCommand(),
     registerSwitchDialectCommand(),
     registerFormatCurrentFileCommand(),
+    registerFormatSqlFilesInDirectoryCommand(),
     registerInsertOrUpdateMetadataHeaderCommand(),
     registerApplySqlovelyExtrasCommand()
   );